Fix for broken test.
Issue was that nothing was actually using the `javax.annotation` module in the project, i.e. a bug in the test project.  For some reason updating the Kotlin jars caused the issue to be displayed.  Fixed by having a module depend on the module. Also added a use of one of the types from one of the patched jars to ensure it working.